+++ tests/test_naming.py
@@ -21,6 +21,8 @@ from flint.naming import (
     create_fits_mask_names,
     create_image_cube_name,
     create_imaging_name_prefix,
+    create_linmos_base_path,
+    create_linmos_names,
     create_ms_name,
     create_name_from_common_fields,
     extract_beam_from_name,
@@ -35,6 +37,7 @@ from flint.naming import (
     raw_ms_format,
     rename_linear_to_stokes,
     split_images,
+    update_beam_resolution_field_in_path,
 )
 
 
@@ -150,6 +153,32 @@ def test_get_beam_resolution_str():
 
     with pytest.raises(ValueError):
         _ = get_beam_resolution_str("Jack")
+
+
+def test_update_beam_resolution_mode_in_path():
+    """Given a path that has a known beam resolution mode in it, update to another"""
+
+    example = Path("SB57516.RACS_0929-81.round4.i.optimal.round4.residual.linmos.fits")
+    expected = Path("SB57516.RACS_0929-81.round4.i.fixed.round4.residual.linmos.fits")
+
+    assert expected == update_beam_resolution_field_in_path(
+        path=example, original_mode="optimal", updated_mode="fixed"
+    )
+    assert expected == update_beam_resolution_field_in_path(
+        path=example, original_mode="optimal", updated_mode="fixed", marker="."
+    )
+    with pytest.raises(AssertionError):
+        update_beam_resolution_field_in_path(
+            path=example, original_mode="fixed", updated_mode="optimal"
+        )
+        assert expected == update_beam_resolution_field_in_path(
+            path=example, original_mode="optimal", updated_mode="fixed", marker="!"
+        )
+        update_beam_resolution_field_in_path(
+            path=Path("JackSparrowCaresNotForBeamResolutions"),
+            original_mode="optimal",
+            updated_mode="fixed",
+        )
 
 
 def test_casda_ms_format_1934():
@@ -733,6 +762,72 @@ def test_create_name_from_common_fields_2():
 
     with pytest.raises(ValueError):
         create_name_from_common_fields(in_paths=examples)
+
+
+def test_create_linmos_parset_base_path():
+    """The yandasoft linmos task writes out a configuration file.
+    This function tests the generation of the path"""
+    examples = get_lots_of_names_2()
+
+    expected = Path("59058/SB59058.RACS_1626-84.round4.i").absolute()
+    assert expected == create_linmos_base_path(input_images=examples)
+
+    expected = Path("59058/SB59058.RACS_1626-84.round4.i.jack.sparrow").absolute()
+    assert expected == create_linmos_base_path(
+        input_images=examples, additional_suffixes="jack.sparrow"
+    )
+    new_paths = [Path("/Here/Be/Pirates") / p for p in examples]
+    expected = Path("/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i").absolute()
+    assert expected == create_linmos_base_path(input_images=new_paths)
+
+    expected = Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.jack.sparrow"
+    ).absolute()
+    assert expected == create_linmos_base_path(
+        input_images=new_paths, additional_suffixes="jack.sparrow"
+    )
+
+
+def test_create_linmos_names():
+    """Some logic is around creating names to give to linmos to create the output images, weights and parset"""
+
+    linmos_names = create_linmos_names(
+        name_prefix="/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i"
+    )
+    assert linmos_names.image_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.linmos.fits"
+    )
+    assert linmos_names.weight_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.weight.fits"
+    )
+    assert linmos_names.parset_output_path == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i_parset.txt"
+    )
+
+    linmos_names = create_linmos_names(
+        name_prefix=Path("/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i")
+    )
+    assert linmos_names.image_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.linmos.fits"
+    )
+    assert linmos_names.weight_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.weight.fits"
+    )
+    assert linmos_names.parset_output_path == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i_parset.txt"
+    )
+
+    linmos_names = create_linmos_names(
+        name_prefix="/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i",
+        parset_output_path=Path("Jack/My/Boi.txt"),
+    )
+    assert linmos_names.image_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.linmos.fits"
+    )
+    assert linmos_names.weight_fits == Path(
+        "/Here/Be/Pirates/59058/SB59058.RACS_1626-84.round4.i.weight.fits"
+    )
+    assert linmos_names.parset_output_path == Path("Jack/My/Boi.txt")
 
 
 def test_rename_linear_to_stokes():